# docs/PickDo.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PickDo",
    "description": "자취생 음식물 낭비 분석 앱 PickDo의 API 명세서",
    "version": "1.0.0"
  },
...
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "로그인",
        "deprecated": false,
        "description": "",
        "tags": [
          "user"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              },
              "example": {
                "email": "user@example.com",
                "password": "P@ssw0rd123"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string"
                        },
                        "refreshToken": {
                          "type": "string"
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer"
                            },
                            "email": {
                              "type": "string"
                            },
                            "displayName": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "email",
                            "displayName"
                          ]
                        }
                      },
                      "required": [
                        "accessToken",
                        "refreshToken",
                        "user"
                      ]
                    },
                    "error": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "error"
                  ]
                },
                "example": {
                  "success": true,
                  "data": {
                    "accessToken": "eyJhbGciOiJIUzI1NiJ9.example",
                    "refreshToken": "eyJhbGciOiJIUzI1NiJ9.refresh",
                    "user": {
                      "id": 1,
                      "email": "user@example.com",
                      "displayName": "홍길동"
                    }
                  },
                  "error": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/auth/register": {
      "post": {
        "summary": "회원가입",
        "deprecated": false,
        "description": "이메일/비밀번호로 신규 계정을 생성한다. 비밀번호는 BCrypt로 users.password_hash에 저장하며 가입 시 토큰은 발급하지 않는다. 이메일은 로그인과 동일하게 대소문자를 구분하며 탈퇴 계정을 포함한 기존 이메일은 재사용할 수 없다.",
        "tags": [
          "user"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "displayName"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 255,
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "writeOnly": true,
                    "maxLength": 72,
                    "description": "공백만으로 구성할 수 없으며 UTF-8 기준 최대 72바이트.",
                    "example": "P@ssw0rd123"
                  },
                  "displayName": {
                    "type": "string",
                    "maxLength": 80,
                    "description": "공백만으로 구성할 수 없다.",
                    "example": "홍길동"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "userId": {
                          "type": "integer"
                        },
                        "email": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        }
                      }
                    },
                    "error": {
                      "type": "object",
                      "properties": {},
                      "nullable": true
                    }
                  }
                },
                "example": {
                  "success": true,
                  "data": {
                    "userId": 12,
                    "email": "user@example.com",
                    "displayName": "홍길동"
                  },
                  "error": null
                }
              }
            },
            "headers": {}
          },
          "409": {
            "description": "이메일 중복 (탈퇴 계정 포함)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "필수값 누락, 이메일 형식 오류 또는 입력 길이 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/auth/{provider}/login": {
      "post": {
        "summary": "소셜 로그인 연동",
        "deprecated": false,
        "description": "외부 인증 제공자(카카오, 구글 등) 로그인 결과로 받은 사용자 식별값으로 로그인하거나 신규 계정을 연동 생성한다.",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "description": "auth_providers.code 값 (예: kakao, google, apple)",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "example": "kakao"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "providerUserKey"
                ],
                "properties": {
                  "providerUserKey": {
                    "type": "string",
                    "example": "kakao_9912345"
                  },
                  "email": {
                    "type": "string",
                    "example": "user@example.com"
                  },
                  "displayName": {
                    "type": "string",
                    "example": "홍길동"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "success": true,
                  "data": {
                    "userId": 12,
                    "accessToken": "eyJhbGciOiJIUzI1NiJ9.access",
                    "refreshToken": "eyJhbGciOiJIUzI1NiJ9.refresh",
                    "isNewUser": false
                  },
                  "error": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorObject": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "INVALID_REQUEST"
          },
          "message": {
            "type": "string",
            "example": "요청이 올바르지 않습니다."
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "data": {
            "type": "null"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorObject"
          }
        },
        "required": [
          "success",
          "error"
        ]
      }
    },
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [],
  "security": []
}